+++ examples/advanced.js
@@ -1,7 +1,7 @@
 var scraper = require('scraper');
 
 scraper({
-	'uri': 'http://search.twitter.com/search?q=nodejs'
+	'uri': 'https://example.test/search?q=nodejs'
 	, 'headers': {
 		'User-Agent': 'Mozilla/4.0 (compatible; MSIE 7.0; Windows NT 6.0)'
 	}}