+++ examples/simple.js
@@ -1,6 +1,6 @@
 var scraper = require('scraper');
 
-scraper('http://search.twitter.com/search?q=javascript', function(err, $) {
+scraper('https://example.test/search?q=javascript', function(err, $) {
 	if (err) {throw err;}
 
 	$('.msg').each(function() {